Move `avatar_list` to index handler
